UI changes and other enhancement in Calendar page

+++ src/app/component/admin/admin.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { AuthService } from 'src/app/services/auth.service';
 import { TranslateService } from '@ngx-translate/core';
+import { CommonService } from 'src/app/services/common.service';
 
 @Component({
   selector: 'app-admin',
@@ -10,14 +11,14 @@ import { TranslateService } from '@ngx-translate/core';
 })
 export class AdminComponent {
 
-  constructor(private router: Router, private authService: AuthService, private translate: TranslateService) { }
+  constructor(private router: Router, private authService: AuthService, private translate: TranslateService,private common:CommonService) { }
 
   logout() {
     const role_id:any = this.authService.roleId ? this.authService.roleId : JSON.parse(<any>localStorage.getItem("roleId"));
     const userId:any = this.authService.userId ? this.authService.userId : JSON.parse(<any>localStorage.getItem("userId"));
     this.authService.logout(role_id, userId, (res: any) => {
-      if (res) {
-        this.router.navigate(['/login']);
+      if (res && localStorage.getItem("isadmin")) {
+        this.router.navigate(['/adminLogin']);
         localStorage.clear();
       } else {
         this.router.navigate(['/login']);
@@ -25,5 +26,11 @@ export class AdminComponent {
       }
     })
 
+  }
+
+  ngOnInit(){
+    this.common.bearerToken.next();
+    this.common.refreshToken.next();
+    
   }
 }
